Draft python event class
