+++ components/function-runtimes/python38/kubeless/requirements.txt
@@ -1,3 +1,4 @@
+requests
 bottle==0.12.13
 cherrypy==8.9.1
 wsgi-request-logger==0.4.6
